+++ yash-env/src/system/shared.rs
@@ -46,6 +46,7 @@ use crate::Env;
 use crate::io::Fd;
 use crate::job::Pid;
 use crate::job::ProcessState;
+use crate::semantics::ExitStatus;
 use enumset::EnumSet;
 use std::cell::RefCell;
 use std::convert::Infallible;
@@ -442,6 +443,9 @@ impl System for &SharedSystem {
     fn execve(&mut self, path: &CStr, args: &[CString], envs: &[CString]) -> Result<Infallible> {
         self.0.borrow_mut().execve(path, args, envs)
     }
+    fn exit(&mut self, exit_status: ExitStatus) -> Pin<Box<dyn Future<Output = Infallible>>> {
+        self.0.borrow_mut().exit(exit_status)
+    }
     fn getcwd(&self) -> Result<PathBuf> {
         self.0.borrow().getcwd()
     }
@@ -664,6 +668,10 @@ impl System for SharedSystem {
     #[inline]
     fn execve(&mut self, path: &CStr, args: &[CString], envs: &[CString]) -> Result<Infallible> {
         (&mut &*self).execve(path, args, envs)
+    }
+    #[inline]
+    fn exit(&mut self, exit_status: ExitStatus) -> Pin<Box<dyn Future<Output = Infallible>>> {
+        (&mut &*self).exit(exit_status)
     }
     #[inline]
     fn getcwd(&self) -> Result<PathBuf> {